requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.10.3
pandas==2.2.3
plotly==5.24.1
streamlit==1.40.2
pyspark==3.5.3
scikit-learn==1.5.2
pytest==8.3.4
httpx==0.28.1
